Bug 1456394 - Moving timing updates to KeyframeEffectReadOnly; r=hiro
This might seem a bit odd but later in this patch series we will rename KeyframeEffectReadOnly to KeyframeEffect.
